rearrange connectivity code
